+++ api/src/main/java/ar/edu/itba/pod/models/dto/TypePercentageByStreetDTO.java
@@ -1,6 +1,8 @@
 package ar.edu.itba.pod.models.dto;
 
+import java.math.BigDecimal;
+
 public record TypePercentageByStreetDTO (
-        String street, double percentage
+        String street, BigDecimal percentage
 ) {
 }
